dom/animation/test/css-animations/test_animation-target.html

<!doctype html>
<meta charset=utf-8>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<div id="log"></div>
<style>
@keyframes anim { }
</style>
<script>
'use strict';

test(function(t) {
  var div = addDiv(t);
  div.style.animation = 'anim 100s';
  var players = div.getAnimationPlayers();
  assert_equals(players[0].source.target, div,
    'Animation.target is the animatable div');
}, 'Returned CSS animations have the correct Animation.target');

</script>
